Created debounce function and added to button class for all user buttons

+++ lib/button/Button.h
@@ -8,12 +8,11 @@ class Button
 private:
   Timer debounceTimer; 
   DigitalIn input; 
-  uint8_t debounceTime;
+  uint64_t debounceTime;
 public:
   Button(PinName pin, PinMode mode, uint8_t debounceTime);     //constructor
   bool debounceRead (bool initialbutton);
   bool read();
-  
 };
 
 #endif  // BUTTON_H
